[feature][CDC base] Supports string type shard fields
